+++ lib/builders/error_builder.rb
