+++ lib/builders/program_builder.rb
@@ -1,0 +1,11 @@
+module Gobstones
+  class ProgramBuilder
+    def initialize(example)
+      @example = example
+    end
+
+    def build(code)
+      return code
+    end
+  end
+end
