+++ lib/gobstones_runner.rb
@@ -10,10 +10,11 @@ Mumukit.configure do |config|
   config.structured = true
 end
 
+require_relative './builders/error_builder'
+require_relative './builders/program_builder'
 require_relative './extensions/string'
+require_relative './render/html_renderer'
 require_relative './test_hook'
 require_relative './metadata_hook'
 require_relative './checker'
 require_relative './multiple_executions_runner'
-require_relative './html_renderer'
-require_relative './error_builder'
