+++ lib/render/boards.html.erb
@@ -3,7 +3,7 @@
 </script>
 
 <script>
-  <%= File.read("lib/dynamic_load_html.js") %>
+  <%= File.read("lib/render/dynamic_load_html.js") %>
 </script>
 
 <script>
